api/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.18

# AI and ML dependencies
openai==1.77.0
pydantic==2.11.4

# LangChain ecosystem
langchain==0.3.1
langchain-openai==0.2.8
langchain-community==0.3.1
langchain-core==0.3.10
langchain-qdrant==0.2.0
langchain-cohere==0.3.0

# LangGraph for multi-agent workflows
langgraph==0.2.45

# Vector database
qdrant-client==1.12.1

# External API integrations
tavily-python==0.5.0
arxiv==2.1.3
cohere==5.11.4

# Data processing
pandas==2.2.3
numpy==2.1.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0

# Optional: RAGAS for evaluation (if needed)
# ragas==0.2.15

# Development dependencies (optional)
# pytest==8.3.4
# black==24.10.0
# isort==5.13.2
